Extract resolvers and migrations
