+++ src/Pckg/Generic/Migration/CreateGenericTablesRequirements.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Pckg\Generic\Migration;
+
+use Pckg\Auth\Migration\CreateAuthTables;
+use Pckg\Migration\Migration;
+use Pckg\Translator\Migration\CreateTranslationsTable;
+
+class CreateGenericTablesRequirements extends Migration
+{
+
+    /**
+     * Those packets need to be installed before this migration.
+     */
+    public function dependencies()
+    {
+        return [
+            // translatable, permissionable
+            CreateLanguagesTable::class,
+            CreateMenuTables::class,
+            CreateListTables::class,
+            CreateTranslationsTable::class,
+            CreateDataAttributesTables::class,
+        ];
+    }
+
+    public function partials()
+    {
+        return [
+            (new CreateSettingsTable())->setRepository($this->repository),
+            (new CreateAuthTables())->setRepository($this->repository),
+        ];
+    }
+}
