--- src/Pckg/Generic/Resolver/ActionsMorph.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Pckg\Generic\Resolver;
-
-use Pckg\Framework\Provider\RouteResolver;
-use CommsCenter\Pagebuilder\Entity\ActionsMorphs;
-
-class ActionsMorph implements RouteResolver
-{
-
-    public function resolve($value)
-    {
-        return (new ActionsMorphs())->where('id', $value)->oneOrFail();
-    }
-
-    public function parametrize($record)
-    {
-        return $record->id;
-    }
-}
